Chess game
For this assignment you will explore how to apply the design ideas
learned in class to design and implement a board game.

You will implement the game of Chess for two players. Your program, when
launched, should draw the board using the terminal and prompt whomever's
turn it is (white or black) for a move. Once the move is executed, the
move should be played and the new board drawn, and the other player
queried.